refactor(ultimate-browsing): split fetch chain modules
