+++ packages/shared-skills/skills/ultimate-browsing/engine/__init__.py
@@ -7,7 +7,8 @@ observations, never to code. See `../SKILL.md` for the No-Site-Name Rule.
 from .validators import Verdict, ValidationResult, validate, CHALLENGE_MARKERS
 from .waf_detector import detect
 from .url_transforms import TRANSFORMS, apply_transform
-from .fetch_chain import fetch, FetchResult, Attempt
+from .fetch_chain import fetch
+from .result_schema import Attempt, FetchResult
 
 __all__ = [
     "Verdict",
