+++ packages/shared-skills/skills/ultimate-browsing/engine/fetch_chain.py
@@ -21,138 +21,14 @@ import json
 import os
 import random
 import time
-from dataclasses import dataclass, field, asdict
-from typing import Any, Optional
+from typing import Optional
 
-from .validators import Verdict, ValidationResult, validate
+from .curl_probe import run_attempt
+from .result_schema import Attempt, FetchResult
+from .summary import format_summary
+from .validators import Verdict
 from .waf_detector import detect, load_profile, _load_profiles, last_load_error
 from .url_transforms import iter_transformed
-
-
-# --- Referer strategies (name → function of original URL) --------------------
-def _self_root(url: str) -> str:
-    from urllib.parse import urlsplit
-    p = urlsplit(url)
-    return f"{p.scheme}://{p.netloc}/"
-
-
-REFERER_STRATEGIES = {
-    "self_root": _self_root,
-    "google_search": lambda _url: "https://www.google.com/",
-    "none": lambda _url: "",
-}
-
-
-# --- Attempt & result schema (Codex: "evidence schema first") ----------------
-@dataclass
-class Attempt:
-    phase: str                       # probe | grid | fallback
-    executor: str                    # curl_cffi | playwright_mcp | playwright_real_chrome | ...
-    url: str
-    url_transform: str               # original | mobile_subdomain | ...
-    impersonate: Optional[str]       # safari | chrome | ... | None (non-curl)
-    referer: str
-    status: int = 0
-    body_size: int = 0
-    verdict: str = ""
-    reasons: list[str] = field(default_factory=list)
-    elapsed_s: float = 0.0
-    error: Optional[str] = None
-
-    def to_dict(self) -> dict:
-        return asdict(self)
-
-
-@dataclass
-class FetchResult:
-    ok: bool
-    content: str = ""
-    final_url: str = ""
-    verdict: str = ""
-    profile_used: Optional[str] = None
-    trace: list[Attempt] = field(default_factory=list)
-    summary: str = ""
-
-    def to_dict(self) -> dict:
-        return {
-            "ok": self.ok,
-            "final_url": self.final_url,
-            "verdict": self.verdict,
-            "profile_used": self.profile_used,
-            "trace": [a.to_dict() for a in self.trace],
-            "summary": self.summary,
-            "content_length": len(self.content),
-        }
-
-
-# --- curl_cffi probe executor ------------------------------------------------
-def _curl_probe(
-    url: str, *, impersonate: str, referer: str, timeout: int = 20
-) -> tuple[Any, Optional[str]]:
-    """Returns (response, error_str). response may be None on exception."""
-    try:
-        from curl_cffi import requests as cffi_requests
-    except ImportError:
-        return None, "curl_cffi not installed"
-
-    headers = {
-        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
-        "Accept-Language": "ko-KR,ko;q=0.9,en-US;q=0.8,en;q=0.7",
-    }
-    if referer:
-        headers["Referer"] = referer
-
-    try:
-        resp = cffi_requests.get(
-            url,
-            impersonate=impersonate,
-            headers=headers,
-            timeout=timeout,
-            allow_redirects=True,
-        )
-        return resp, None
-    except Exception as e:
-        return None, f"{type(e).__name__}:{str(e)[:200]}"
-
-
-def _run_attempt(
-    url: str,
-    *,
-    transform_name: str,
-    impersonate: str,
-    referer_name: str,
-    success_selectors: Optional[list[str]],
-    known_bad_sizes: Optional[list[int]],
-    timeout: int,
-    phase: str,
-) -> tuple[Attempt, Any]:
-    """Execute one curl_cffi attempt and produce an Attempt record."""
-    referer_url = REFERER_STRATEGIES.get(referer_name, REFERER_STRATEGIES["none"])(url)
-    t0 = time.time()
-    resp, err = _curl_probe(url, impersonate=impersonate, referer=referer_url, timeout=timeout)
-    elapsed = round(time.time() - t0, 3)
-
-    att = Attempt(
-        phase=phase,
-        executor="curl_cffi",
-        url=url,
-        url_transform=transform_name,
-        impersonate=impersonate,
-        referer=referer_name,
-        elapsed_s=elapsed,
-    )
-
-    if err or resp is None:
-        att.error = err or "no response"
-        att.verdict = Verdict.UNKNOWN.value
-        return att, None
-
-    vr = validate(resp, success_selectors=success_selectors, known_bad_sizes=known_bad_sizes)
-    att.status = vr.status
-    att.body_size = vr.body_size
-    att.verdict = vr.verdict.value
-    att.reasons = vr.reasons
-    return att, resp
 
 
 # --- Main entrypoint ---------------------------------------------------------
@@ -218,7 +94,7 @@ def fetch(
     if device_class == "mobile":
         base_impersonate = user_hint.get("impersonate_first") or "safari_ios"
 
-    probe_attempt, probe_resp = _run_attempt(
+    probe_attempt, probe_resp = run_attempt(
         url,
         transform_name="original",
         impersonate=base_impersonate,
@@ -277,7 +153,7 @@ def fetch(
                     if (t_name == "original" and tls == base_impersonate
                             and ref == (user_hint.get("referer_strategy") or "self_root")):
                         continue
-                    att, resp = _run_attempt(
+                    att, resp = run_attempt(
                         t_url,
                         transform_name=t_name,
                         impersonate=tls,
@@ -369,7 +245,7 @@ def fetch(
             ))
 
     # -------- Give up, return best we have ----------------------------------
-    summary = _format_summary(trace, profile_used)
+    summary = format_summary(trace, profile_used)
     return FetchResult(
         ok=False,
         content=getattr(last_resp, "text", "") if last_resp is not None else "",
@@ -391,38 +267,3 @@ def _build_result(resp, attempt: Attempt, trace: list[Attempt], profile_used: Op
         trace=trace,
         summary=f"{attempt.executor} {attempt.impersonate} + {attempt.url_transform} + referer:{attempt.referer} → {attempt.verdict}",
     )
-
-
-# WAF profiles known to typically gate HTML but leave internal JSON APIs
-# (relatively) open. When these are detected and curl challenges pile up,
-# we surface R7 hint in the summary so the caller (or Claude) can branch
-# to an API-first route without waiting for full grid exhaustion.
-_R7_ELIGIBLE_PROFILES = frozenset({
-    "akamai_bot_manager",
-    "cloudflare_turnstile",
-    "datadome_probable",
-    "perimeterx_human",
-    "f5_big_ip",
-    "aws_waf",
-})
-
-R7_HINT = (
-    "💡 R7 API-first 권장: WAF가 HTML 경로를 차단 중. "
-    "Playwright MCP 사용 → browser_navigate → browser_network_requests "
-    "→ `/api/`·`/graphql`·`\\.json` 필터로 내부 엔드포인트 탐지 → "
-    "해당 URL을 `python3 -m engine <API_URL>`로 재호출. 대부분 API 레이어는 "
-    "WAF 방어가 얕아 curl_cffi만으로 수집됨."
-)
-
-
-def _format_summary(trace: list[Attempt], profile: Optional[str]) -> str:
-    n = len(trace)
-    verdicts = [a.verdict for a in trace]
-    challenge_count = sum(1 for v in verdicts if v == Verdict.CHALLENGE.value)
-    base = (
-        f"failed after {n} attempts; profile={profile}; "
-        f"verdicts={','.join(v for v in verdicts[:5])}" + ("..." if n > 5 else "")
-    )
-    if profile in _R7_ELIGIBLE_PROFILES and challenge_count >= 3:
-        return base + "\n" + R7_HINT
-    return base
