+++ packages/shared-skills/skills/ultimate-browsing/engine/referers.py
@@ -1,0 +1,15 @@
+from __future__ import annotations
+
+from urllib.parse import urlsplit
+
+
+def _self_root(url: str) -> str:
+    parsed = urlsplit(url)
+    return f"{parsed.scheme}://{parsed.netloc}/"
+
+
+REFERER_STRATEGIES = {
+    "self_root": _self_root,
+    "google_search": lambda _url: "https://www.google.com/",
+    "none": lambda _url: "",
+}
